requirements.txt
azure_storage==0.37.0
Django==3.2.4
matplotlib==3.4.3
msrest==0.6.21
numpy==1.22.3
opencv_python==4.5.3.56
pandas==1.2.5
Pillow==9.0.1
requests==2.25.1
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.7.0
seaborn==0.11.2
skimage==0.0
tensorflow==2.6.0
